refactor: Store component includes instance of Main Store
